# docs/01_contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Exhibition Infrastructure API
  version: 0.1.0
  description: >
    Phase 2A OpenAPI skeleton for Flutter App canonical paths, plus retained Admin and
    internal service alignment placeholders.
servers:
  - url: http://localhost:3000
    description: Local BFF
  - url: http://localhost:3001
    description: Local Server
...
paths:
...
  /api/app/auth/otp/send:
    post:
      tags: [Auth]
      summary: OTP send canonical path for Flutter App login
      operationId: postAppAuthOtpSend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpSendRequest'
      responses:
        "200":
          description: Minimum OTP send acknowledgment for Flutter App.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpSendResponse'
  /api/app/auth/otp/login:
    post:
      tags: [Auth]
      summary: OTP login canonical path for Flutter App
      operationId: postAppAuthOtpLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpLoginRequest'
      responses:
        "200":
          description: Minimum session establish response for Flutter App.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEstablishResponse'
  /api/app/auth/refresh:
    post:
      tags: [Auth]
      summary: Session refresh canonical path for Flutter App
      operationId: postAppAuthRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshSessionRequest'
      responses:
        "200":
          description: Rotated app-facing access and refresh token projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionRefreshResponse'
  /api/app/auth/logout:
    post:
      tags: [Auth]
      summary: Logout canonical path for Flutter App
      operationId: postAppAuthLogout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        "200":
          description: Minimum logout acknowledgment for Flutter App.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionAckResponse'
...
  /api/app/exhibition/enterprise-hub/enterprises:
    get:
      tags: [Exhibition]
      summary: Enterprise hub board list canonical path for Flutter App
      operationId: getAppEnterpriseHubEnterprises
      parameters:
        - in: query
          name: boardType
          required: true
          schema:
            $ref: '#/components/schemas/EnterpriseHubBoardType'
        - in: query
          name: keyword
          required: false
          schema:
            type: string
        - in: query
          name: provinceCode
          required: false
          schema:
            type: string
        - in: query
          name: cityCode
          required: false
          schema:
            type: string
        - in: query
          name: certifiedOnly
          required: false
          schema:
            type: boolean
        - in: query
          name: sortBy
          required: false
          schema:
            type: string
            enum:
              - default
              - latest
              - cases_desc
        - in: query
          name: exhibitionType
          required: false
          schema:
            type: string
        - in: query
          name: serviceCity
          required: false
          schema:
            type: string
        - in: query
          name: caseCountRange
          required: false
          schema:
            type: string
        - in: query
          name: reputationLevel
          required: false
          schema:
            type: string
        - in: query
          name: processType
          required: false
          schema:
            type: string
        - in: query
          name: plantAreaRange
          required: false
          schema:
            type: string
        - in: query
          name: urgentCapability
          required: false
          schema:
            $ref: '#/components/schemas/EnterpriseHubUrgentOrderCapability'
        - in: query
          name: warehouseCapability
          required: false
          schema:
            type: boolean
        - in: query
          name: supplyCategory
          required: false
          schema:
            type: string
        - in: query
          name: supplyMode
          required: false
          schema:
            type: string
        - in: query
          name: responseLevel
          required: false
          schema:
            type: string
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
        - in: query
          name: pageSize
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: >
            Enterprise hub board list projection for Flutter App. This path
            freezes one list read model only and must not become organization
            identity truth, certification truth, or a second ranking state
            machine.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnterpriseHubListResponse'
        "400":
          description: >
            Controlled invalid board or query response. `BFF` must return
            `ENTERPRISE_HUB_INVALID_BOARD_TYPE` when boardType is missing or
            unsupported.
  /api/app/exhibition/enterprise-hub/enterprises/{enterpriseId}:
    get:
      tags: [Exhibition]
      summary: Enterprise hub detail canonical path for Flutter App
      operationId: getAppEnterpriseHubEnterpriseDetail
      parameters:
        - in: path
          name: enterpriseId
          required: true
          schema:
            type: string
        - in: query
          name: boardType
          required: true
          schema:
            $ref: '#/components/schemas/EnterpriseHubBoardType'
      responses:
        "200":
          description: >
            Enterprise hub detail read model for Flutter App. This path
            freezes one details projection only and must not become
            organization, certification, or recommendation truth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnterpriseHubDetailResponse'
        "404":
          description: >
            Controlled enterprise-unavailable response. `Server` must return
            `ENTERPRISE_HUB_ENTERPRISE_NOT_FOUND` when the requested listing is
            absent or hidden under the current boundary.
  /api/app/exhibition/enterprise-hub/recommendations:
    get:
      tags: [Exhibition]
      summary: Enterprise hub recommendation slots canonical path for Flutter App
      operationId: getAppEnterpriseHubRecommendations
      parameters:
        - in: query
          name: boardType
          required: true
          schema:
            $ref: '#/components/schemas/EnterpriseHubBoardType'
      responses:
        "200":
          description: >
            Enterprise hub recommendation slot projection for the selected
            board only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnterpriseHubRecommendationListResponse'
  /api/app/exhibition/enterprise-hub/applications:
    post:
      tags: [Exhibition]
      summary: Create enterprise hub application canonical path for Flutter App
      operationId: createAppEnterpriseHubApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseHubCreateApplicationRequest'
      responses:
        "200":
          description: Created current enterprise hub application draft under current organization scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnterpriseHubCreateApplicationResponse'
        "400":
          description: >
            Controlled application-create invalid response. `Server` must
            return `ENTERPRISE_HUB_MISSING_REQUIRED_FIELDS` when the minimum
            create boundary is not met.
  /api/app/exhibition/enterprise-hub/enterprises/{enterpriseId}/basic:
    put:
      tags: [Exhibition]
      summary: Update enterprise hub basic listing profile canonical path for Flutter App
      operationId: updateAppEnterpriseHubBasic
      parameters:
        - in: path
          name: enterpriseId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseHubUpdateBasicRequest'
      responses:
        "200":
          description: Enterprise hub basic listing profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionAckResponse'
  /api/app/exhibition/enterprise-hub/enterprises/{enterpriseId}/profiles/company:
    put:
      tags: [Exhibition]
      summary: Update enterprise hub company profile canonical path for Flutter App
      operationId: updateAppEnterpriseHubCompanyProfile
      parameters:
        - in: path
          name: enterpriseId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseHubUpdateCompanyProfileRequest'
      responses:
        "200":
          description: Enterprise hub company profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionAckResponse'
  /api/app/exhibition/enterprise-hub/enterprises/{enterpriseId}/profiles/factory:
    put:
      tags: [Exhibition]
      summary: Update enterprise hub factory profile canonical path for Flutter App
      operationId: updateAppEnterpriseHubFactoryProfile
      parameters:
        - in: path
          name: enterpriseId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseHubUpdateFactoryProfileRequest'
      responses:
        "200":
          description: Enterprise hub factory profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionAckResponse'
  /api/app/exhibition/enterprise-hub/enterprises/{enterpriseId}/profiles/supplier:
    put:
      tags: [Exhibition]
      summary: Update enterprise hub supplier profile canonical path for Flutter App
      operationId: updateAppEnterpriseHubSupplierProfile
      parameters:
        - in: path
          name: enterpriseId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseHubUpdateSupplierProfileRequest'
      responses:
        "200":
          description: Enterprise hub supplier profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionAckResponse'
  /api/app/exhibition/enterprise-hub/enterprises/{enterpriseId}/cases:
    post:
      tags: [Exhibition]
      summary: Create enterprise hub case canonical path for Flutter App
      operationId: createAppEnterpriseHubCase
      parameters:
        - in: path
          name: enterpriseId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseHubCreateCaseRequest'
      responses:
        "200":
          description: Enterprise hub case created in draft or pending-review state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnterpriseHubCaseCreateResponse'
  /api/app/exhibition/enterprise-hub/applications/{applicationId}/submit:
    post:
      tags: [Exhibition]
      summary: Submit enterprise hub application canonical path for Flutter App
      operationId: submitAppEnterpriseHubApplication
      parameters:
        - in: path
          name: applicationId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseHubSubmitApplicationRequest'
      responses:
        "200":
          description: Enterprise hub application submitted into current review flow.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionAckResponse'
        "400":
          description: >
            Controlled submit invalid response. `Server` must return one of
            `ENTERPRISE_HUB_PROFILE_NOT_COMPLETED`,
            `ENTERPRISE_HUB_CERTIFICATION_REQUIRED`,
            `ENTERPRISE_HUB_CONTACT_REQUIRED`, or
            `ENTERPRISE_HUB_CASE_REQUIRED` when the current submit boundary is
            not met.
  /api/app/exhibition/enterprise-hub/applications/{applicationId}:
    get:
      tags: [Exhibition]
      summary: Enterprise hub application status canonical path for Flutter App
      operationId: getAppEnterpriseHubApplicationStatus
      parameters:
        - in: path
          name: applicationId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current enterprise hub application status projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnterpriseHubApplicationStatusResponse'
        "404":
          description: >
            Controlled application-unavailable response. `Server` must return
            `ENTERPRISE_HUB_APPLICATION_NOT_FOUND` when the current
            application is absent or out of actor scope.
...
  /server/admin/exhibition/enterprise-hub/applications:
    get:
      tags: [Admin]
      summary: Admin enterprise hub application list canonical path
      operationId: getAdminEnterpriseHubApplications
      parameters:
        - in: query
          name: applicationStatus
          required: false
          schema:
            $ref: '#/components/schemas/EnterpriseHubApplicationStatus'
        - in: query
          name: boardType
          required: false
          schema:
            $ref: '#/components/schemas/EnterpriseHubBoardType'
        - in: query
          name: provinceCode
          required: false
          schema:
            type: string
        - in: query
          name: keyword
          required: false
          schema:
            type: string
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
        - in: query
          name: pageSize
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Enterprise hub application list for controlled admin review only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnterpriseHubAdminApplicationListResponse'
  /server/admin/exhibition/enterprise-hub/applications/{applicationId}:
    get:
      tags: [Admin]
      summary: Admin enterprise hub application detail canonical path
      operationId: getAdminEnterpriseHubApplicationDetail
      parameters:
        - in: path
          name: applicationId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Enterprise hub application detail for controlled admin review only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnterpriseHubAdminApplicationDetailResponse'
...
components:
  schemas:
...
    OtpSendRequest:
      type: object
      required:
        - mobile
        - scene
      properties:
        mobile:
          type: string
        scene:
          type: string
          enum:
            - login
        deviceId:
          type:
            - string
            - 'null'
        captchaToken:
          type:
            - string
            - 'null'
    OtpSendResponse:
      type: object
      required:
        - cooldownSeconds
        - traceId
      properties:
        cooldownSeconds:
          type: integer
          minimum: 0
        traceId:
          type: string
    OtpLoginRequest:
      type: object
      required:
        - mobile
        - otpCode
        - deviceId
      properties:
        mobile:
          type: string
        otpCode:
          type: string
        deviceId:
          type: string
        deviceName:
          type:
            - string
            - 'null'
        osType:
          type:
            - string
            - 'null'
        appVersion:
          type:
            - string
            - 'null'
    SessionEstablishResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresInSeconds
        - shellBootstrapState
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresInSeconds:
          type: integer
          minimum: 1
        shellBootstrapState:
          type: string
          enum:
            - authenticated
            - no_organization
    RefreshSessionRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
        deviceId:
          type:
            - string
            - 'null'
    SessionRefreshResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresInSeconds
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresInSeconds:
          type: integer
          minimum: 1
    LogoutRequest:
      type: object
      properties:
        revokeAllOtherDevices:
          type: boolean
        deviceId:
          type:
            - string
            - 'null'
    ActionAckResponse:
      type: object
      required:
        - ok
        - traceId
      properties:
        ok:
          type: boolean
        traceId:
          type: string
...
    EnterpriseHubBoardType:
      type: string
      enum:
        - company
        - factory
        - supplier
    EnterpriseHubApplicationStatus:
      type: string
      enum:
        - draft
        - submitted
        - under_review
        - revision_required
        - approved
        - rejected
...
    EnterpriseHubTeamSizeRange:
      type: string
      enum:
        - 1_10
        - 11_30
        - 31_100
        - 101_300
        - 300_plus
    EnterpriseHubUrgentOrderCapability:
      type: string
      enum:
        - none
        - 24h
        - 48h
        - 72h
        - custom
...
    EnterpriseHubPagination:
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        hasMore:
          type: boolean
...
    EnterpriseHubListItem:
      type: object
      required:
        - enterpriseId
        - boardType
        - name
        - provinceName
        - cityName
        - primaryBoardLabel
        - secondaryCapabilityLabels
        - shortIntro
        - certificationLabel
        - caseCount
        - boardHighlights
      properties:
        enterpriseId:
          type: string
        boardType:
          $ref: '#/components/schemas/EnterpriseHubBoardType'
        name:
          type: string
        logoUrl:
          type:
            - string
            - 'null'
        provinceName:
          type: string
        cityName:
          type: string
        primaryBoardLabel:
          type: string
        secondaryCapabilityLabels:
          type: array
          items:
            type: string
          uniqueItems: true
        shortIntro:
          type: string
        certificationLabel:
          type: string
        caseCount:
          type: integer
          minimum: 0
        avgScore:
          type:
            - number
            - 'null'
        keywordTags:
          type: array
          items:
            type: string
        boardHighlights:
          $ref: '#/components/schemas/EnterpriseHubBoardHighlights'
    EnterpriseHubListResponse:
      type: object
      required:
        - recommended
        - items
        - pagination
      properties:
        recommended:
          type: array
          items:
            $ref: '#/components/schemas/EnterpriseHubListItem'
        items:
          type: array
          items:
            $ref: '#/components/schemas/EnterpriseHubListItem'
        pagination:
          $ref: '#/components/schemas/EnterpriseHubPagination'
...
    EnterpriseHubBasicInfo:
      type: object
      required:
        - fullIntro
      properties:
        legalName:
          type:
            - string
            - 'null'
        foundedAt:
          type:
            - string
            - 'null'
          format: date
        teamSizeRange:
          oneOf:
            - $ref: '#/components/schemas/EnterpriseHubTeamSizeRange'
            - type: 'null'
        fullIntro:
          type:
            - string
            - 'null'
        address:
          type:
            - string
            - 'null'
...
    EnterpriseHubDetailResponse:
      type: object
      required:
        - header
        - basicInfo
        - boardProfile
        - serviceAreas
        - cases
        - certifications
        - reviewSummary
        - contacts
      properties:
        header:
          $ref: '#/components/schemas/EnterpriseHubHeader'
        basicInfo:
          $ref: '#/components/schemas/EnterpriseHubBasicInfo'
        boardProfile:
          $ref: '#/components/schemas/EnterpriseHubBoardProfile'
        serviceAreas:
          type: array
          items:
            $ref: '#/components/schemas/EnterpriseHubServiceArea'
        cases:
          type: array
          items:
            $ref: '#/components/schemas/EnterpriseHubCaseCard'
        certifications:
          type: array
          items:
            $ref: '#/components/schemas/EnterpriseHubCertificationCard'
        reviewSummary:
          $ref: '#/components/schemas/EnterpriseHubReviewSummary'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/EnterpriseHubContactCard'
    EnterpriseHubRecommendationListResponse:
      type: object
      required:
        - boardType
        - items
      properties:
        boardType:
          $ref: '#/components/schemas/EnterpriseHubBoardType'
        items:
          type: array
          items:
            $ref: '#/components/schemas/EnterpriseHubListItem'
    EnterpriseHubCreateApplicationRequest:
      type: object
      required:
        - applyBoardType
        - applicantName
        - applicantMobile
      description: >
        Create one enterprise hub application draft under the current
        authenticated organization scope only.
      properties:
        applyBoardType:
          $ref: '#/components/schemas/EnterpriseHubBoardType'
        applicantName:
          type: string
        applicantMobile:
          type: string
    EnterpriseHubCreateApplicationResponse:
      type: object
      required:
        - applicationId
        - enterpriseId
        - applicationStatus
      properties:
        applicationId:
          type: string
        enterpriseId:
          type: string
        applicationStatus:
          $ref: '#/components/schemas/EnterpriseHubApplicationStatus'
    EnterpriseHubUpdateBasicRequest:
      type: object
      properties:
        name:
          type:
            - string
            - 'null'
        logoFileAssetId:
          type:
            - string
            - 'null'
        coverFileAssetId:
          type:
            - string
            - 'null'
        shortIntro:
          type:
            - string
            - 'null'
        fullIntro:
          type:
            - string
            - 'null'
        provinceCode:
          type:
            - string
            - 'null'
        provinceName:
          type:
            - string
            - 'null'
        cityCode:
          type:
            - string
            - 'null'
        cityName:
          type:
            - string
            - 'null'
        address:
          type:
            - string
            - 'null'
        foundedAt:
          type:
            - string
            - 'null'
          format: date
        teamSizeRange:
          oneOf:
            - $ref: '#/components/schemas/EnterpriseHubTeamSizeRange'
            - type: 'null'
        cooperationModes:
          type: array
          items:
            type: string
        contactVisible:
          type:
            - boolean
            - 'null'
    EnterpriseHubUpdateCompanyProfileRequest:
      $ref: '#/components/schemas/EnterpriseHubCompanyProfile'
    EnterpriseHubUpdateFactoryProfileRequest:
      $ref: '#/components/schemas/EnterpriseHubFactoryProfile'
    EnterpriseHubUpdateSupplierProfileRequest:
      $ref: '#/components/schemas/EnterpriseHubSupplierProfile'
    EnterpriseHubCreateCaseRequest:
      type: object
      required:
        - boardType
        - title
        - summary
        - caseCoverFileAssetId
      properties:
        boardType:
          $ref: '#/components/schemas/EnterpriseHubBoardType'
        title:
          type: string
        exhibitionType:
          type:
            - string
            - 'null'
        city:
          type:
            - string
            - 'null'
        eventTime:
          type:
            - string
            - 'null'
          format: date
        summary:
          type: string
        caseCoverFileAssetId:
          type: string
        caseMediaFileAssetIds:
          type: array
          items:
            type: string
        isFeatured:
          type:
            - boolean
            - 'null'
    EnterpriseHubCaseCreateResponse:
      type: object
      required:
        - caseId
        - caseStatus
      properties:
        caseId:
          type: string
        caseStatus:
          $ref: '#/components/schemas/EnterpriseHubCaseStatus'
    EnterpriseHubSubmitApplicationRequest:
      type: object
      required:
        - confirm
      properties:
        confirm:
          type: boolean
    EnterpriseHubApplicationStatusResponse:
      type: object
      required:
        - applicationId
        - enterpriseId
        - applyBoardType
        - applicationStatus
      properties:
        applicationId:
          type: string
        enterpriseId:
          type: string
        applyBoardType:
          $ref: '#/components/schemas/EnterpriseHubBoardType'
        applicationStatus:
          $ref: '#/components/schemas/EnterpriseHubApplicationStatus'
        rejectionReason:
          type:
            - string
            - 'null'
        submittedAt:
          type:
            - string
            - 'null'
          format: date-time
        reviewedAt:
          type:
            - string
            - 'null'
          format: date-time
...
    EnterpriseHubAdminApplicationListResponse:
      type: object
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EnterpriseHubAdminApplicationListItem'
        pagination:
          $ref: '#/components/schemas/EnterpriseHubPagination'
...
    EnterpriseHubAdminApplicationDetailResponse:
      type: object
      required:
        - application
        - enterprise
        - profiles
        - cases
        - certifications
        - contacts
      properties:
        application:
          $ref: '#/components/schemas/EnterpriseHubApplicationStatusResponse'
        enterprise:
          $ref: '#/components/schemas/EnterpriseHubAdminEnterpriseSummary'
        profiles:
          $ref: '#/components/schemas/EnterpriseHubAdminProfiles'
        cases:
          type: array
          items:
            $ref: '#/components/schemas/EnterpriseHubCaseCard'
        certifications:
          type: array
          items:
            $ref: '#/components/schemas/EnterpriseHubCertificationCard'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/EnterpriseHubContactCard'
...